refactor: move notification reads to notification/api/notificationReads
